Added function to calculate output and pos differences
The differences interested are euclidean distance and angle differences.
The euclidean distance is basically 3d pythagoras.
The angle is obtained by first getting the transform from output frame to
data provided. Then obtain the quaternion of the transform and from the
'w' component, obtained the angle/2 by using arccos. The angle is then
obtained