Normalised each dimension with their own max and min
It was previously normalised by the max and min in the whole translation coordinates for the positions (e.g. the 3 columns for pos is normalised by max and min in all 3 columns), and in all euler angles (also max and min in the all 3 columns). This is now changed to normalised max and min in each dimension (e.g. max and min in each column).